social-data-service/X_fetcher/requirements.txt
git+https://github.com/tweepy/tweepy.git
python-dotenv==0.19.2
pika==1.3.1
celery==5.2.7
flask==2.0.3
werkzeug==2.0.3
flask
flask-cors==3.0.10  # Optional, for handling CORS
